+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v5.3.7
+----------
+ * Release URNs on Org release
+
+v5.3.6
+----------
+ * Release Channel sync events and alarms
+
+v5.3.5
+----------
+ * release Campaigns when releasing Orgs
+
 v5.3.4
 ----------
  * Release flow starts when releasing flows